% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{classTranslator}
\alias{classTranslator}
\title{Translate a relationship class code on standardized ibdrel format to readable relationships}
\usage{
classTranslator(class, resolution)
}
\description{
Translate a relationship class code on standardized ibdrel format to readable relationships
}
