% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{degree_to_l}
\alias{degree_to_l}
\title{Convert a pedigree path degree to possible path lengths}
\usage{
degree_to_l(degree, full = T)
}
\arguments{
\item{degree}{An integer: The pedigree (path) degree}

\item{full}{Logical: Is the relationship full (TRUE) or half (FALSE)}
}
\value{
A DataFrame
}
\description{
Convert a pedigree path degree to possible path lengths
}
