% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{identify_path_lengths}
\alias{identify_path_lengths}
\title{Convert multiple path degrees to possible path lengths}
\usage{
identify_path_lengths(min_path_degree = 0, max_path_degree)
}
\arguments{
\item{min_path_degree}{An integer: Smallest path degree}

\item{max_path_degree}{An integer: Largest path degree}
}
\value{
A DataFrame
}
\description{
Convert multiple path degrees to possible path lengths
}
