% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Launch the ibdrel app}
\usage{
launchApp()
}
\description{
This launches the Shiny application for performing pairwise relationship estimation.
}
\examples{
\dontrun{
launchApp()
}

}
